# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Universal Annotator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AnnotationType =
  (typeof AnnotationType)[keyof typeof AnnotationType];

export const AnnotationType = {
  text: "text",
  highlight: "highlight",
  drawing: "drawing",
  link: "link",
} as const;

export interface Annotation {
  id: number;
  title: string;
  content: string;
  type: AnnotationType;
  color: string;
  /** @nullable */
  sourceUrl: string | null;
  /** @nullable */
  sourceTitle: string | null;
  isPinned: boolean;
  tags: string[];
  createdAt: string;
  updatedAt: string;
}

export type CreateAnnotationBodyType =
  (typeof CreateAnnotationBodyType)[keyof typeof CreateAnnotationBodyType];

export const CreateAnnotationBodyType = {
  text: "text",
  highlight: "highlight",
  drawing: "drawing",
  link: "link",
} as const;

export interface CreateAnnotationBody {
  title: string;
  content: string;
  type: CreateAnnotationBodyType;
  color?: string;
  /** @nullable */
  sourceUrl?: string | null;
  /** @nullable */
  sourceTitle?: string | null;
  tags?: string[];
}

export type UpdateAnnotationBodyType =
  (typeof UpdateAnnotationBodyType)[keyof typeof UpdateAnnotationBodyType];

export const UpdateAnnotationBodyType = {
  text: "text",
  highlight: "highlight",
  drawing: "drawing",
  link: "link",
} as const;

export interface UpdateAnnotationBody {
  title?: string;
  content?: string;
  type?: UpdateAnnotationBodyType;
  color?: string;
  /** @nullable */
  sourceUrl?: string | null;
  /** @nullable */
  sourceTitle?: string | null;
  tags?: string[];
  isPinned?: boolean;
}

export type AnnotationStatsByType = {
  text: number;
  highlight: number;
  drawing: number;
  link: number;
};

export interface AnnotationStats {
  total: number;
  byType: AnnotationStatsByType;
  pinned: number;
  totalTags: number;
}

export interface Tag {
  id: number;
  name: string;
  color: string;
  annotationCount: number;
  createdAt: string;
}

export interface CreateTagBody {
  name: string;
  color?: string;
}

export type ListAnnotationsParams = {
  /**
   * Filter by tag name
   */
  tag?: string;
  /**
   * Filter by annotation type
   */
  type?: ListAnnotationsType;
  /**
   * Search annotations by content or title
   */
  search?: string;
  /**
   * Filter pinned annotations
   */
  pinned?: ListAnnotationsPinned;
};

export type ListAnnotationsType =
  (typeof ListAnnotationsType)[keyof typeof ListAnnotationsType];

export const ListAnnotationsType = {
  text: "text",
  highlight: "highlight",
  drawing: "drawing",
  link: "link",
} as const;

export type ListAnnotationsPinned =
  (typeof ListAnnotationsPinned)[keyof typeof ListAnnotationsPinned];

export const ListAnnotationsPinned = {
  true: "true",
  false: "false",
} as const;

export type GetRecentAnnotationsParams = {
  /**
   * Max number of results (default 10)
   */
  limit?: number;
};