# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Universal Annotator API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all annotations
 */
export const ListAnnotationsQueryParams = zod.object({
  tag: zod.coerce.string().optional().describe("Filter by tag name"),
  type: zod
    .enum(["text", "highlight", "drawing", "link"])
    .optional()
    .describe("Filter by annotation type"),
  search: zod.coerce
    .string()
    .optional()
    .describe("Search annotations by content or title"),
  pinned: zod
    .enum(["true", "false"])
    .optional()
    .describe("Filter pinned annotations"),
});

export const ListAnnotationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  content: zod.string(),
  type: zod.enum(["text", "highlight", "drawing", "link"]),
  color: zod.string(),
  sourceUrl: zod.string().nullable(),
  sourceTitle: zod.string().nullable(),
  isPinned: zod.boolean(),
  tags: zod.array(zod.string()),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListAnnotationsResponse = zod.array(ListAnnotationsResponseItem);

/**
 * @summary Create a new annotation
 */
export const CreateAnnotationBody = zod.object({
  title: zod.string(),
  content: zod.string(),
  type: zod.enum(["text", "highlight", "drawing", "link"]),
  color: zod.string().optional(),
  sourceUrl: zod.string().nullish(),
  sourceTitle: zod.string().nullish(),
  tags: zod.array(zod.string()).optional(),
});

/**
 * @summary Get annotation statistics
 */
export const GetAnnotationStatsResponse = zod.object({
  total: zod.number(),
  byType: zod.object({
    text: zod.number(),
    highlight: zod.number(),
    drawing: zod.number(),
    link: zod.number(),
  }),
  pinned: zod.number(),
  totalTags: zod.number(),
});

/**
 * @summary Get recently created or modified annotations
 */
export const GetRecentAnnotationsQueryParams = zod.object({
  limit: zod.coerce
    .number()
    .optional()
    .describe("Max number of results (default 10)"),
});

export const GetRecentAnnotationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  content: zod.string(),
  type: zod.enum(["text", "highlight", "drawing", "link"]),
  color: zod.string(),
  sourceUrl: zod.string().nullable(),
  sourceTitle: zod.string().nullable(),
  isPinned: zod.boolean(),
  tags: zod.array(zod.string()),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const GetRecentAnnotationsResponse = zod.array(
  GetRecentAnnotationsResponseItem,
);

/**
 * @summary Get a single annotation
 */
export const GetAnnotationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetAnnotationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  content: zod.string(),
  type: zod.enum(["text", "highlight", "drawing", "link"]),
  color: zod.string(),
  sourceUrl: zod.string().nullable(),
  sourceTitle: zod.string().nullable(),
  isPinned: zod.boolean(),
  tags: zod.array(zod.string()),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update an annotation
 */
export const UpdateAnnotationParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateAnnotationBody = zod.object({
  title: zod.string().optional(),
  content: zod.string().optional(),
  type: zod.enum(["text", "highlight", "drawing", "link"]).optional(),
  color: zod.string().optional(),
  sourceUrl: zod.string().nullish(),
  sourceTitle: zod.string().nullish(),
  tags: zod.array(zod.string()).optional(),
  isPinned: zod.boolean().optional(),
});

export const UpdateAnnotationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  content: zod.string(),
  type: zod.enum(["text", "highlight", "drawing", "link"]),
  color: zod.string(),
  sourceUrl: zod.string().nullable(),
  sourceTitle: zod.string().nullable(),
  isPinned: zod.boolean(),
  tags: zod.array(zod.string()),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete an annotation
 */
export const DeleteAnnotationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Toggle pin status of an annotation
 */
export const ToggleAnnotationPinParams = zod.object({
  id: zod.coerce.number(),
});

export const ToggleAnnotationPinResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  content: zod.string(),
  type: zod.enum(["text", "highlight", "drawing", "link"]),
  color: zod.string(),
  sourceUrl: zod.string().nullable(),
  sourceTitle: zod.string().nullable(),
  isPinned: zod.boolean(),
  tags: zod.array(zod.string()),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary List all tags with annotation counts
 */
export const ListTagsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  color: zod.string(),
  annotationCount: zod.number(),
  createdAt: zod.coerce.date(),
});
export const ListTagsResponse = zod.array(ListTagsResponseItem);

/**
 * @summary Create a tag
 */
export const CreateTagBody = zod.object({
  name: zod.string(),
  color: zod.string().optional(),
});