# lib/api-zod/src/generated/types/annotation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Universal Annotator API
 * OpenAPI spec version: 0.1.0
 */
import type { AnnotationType } from "./annotationType";

export interface Annotation {
  id: number;
  title: string;
  content: string;
  type: AnnotationType;
  color: string;
  /** @nullable */
  sourceUrl: string | null;
  /** @nullable */
  sourceTitle: string | null;
  isPinned: boolean;
  tags: string[];
  createdAt: Date;
  updatedAt: Date;
}