# lib/api-zod/src/generated/types/annotationStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Universal Annotator API
 * OpenAPI spec version: 0.1.0
 */
import type { AnnotationStatsByType } from "./annotationStatsByType";

export interface AnnotationStats {
  total: number;
  byType: AnnotationStatsByType;
  pinned: number;
  totalTags: number;
}