# lib/api-zod/src/generated/types/createAnnotationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Universal Annotator API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAnnotationBodyType } from "./createAnnotationBodyType";

export interface CreateAnnotationBody {
  title: string;
  content: string;
  type: CreateAnnotationBodyType;
  color?: string;
  /** @nullable */
  sourceUrl?: string | null;
  /** @nullable */
  sourceTitle?: string | null;
  tags?: string[];
}