# lib/api-zod/src/generated/types/createTagBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Universal Annotator API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTagBody {
  name: string;
  color?: string;
}