# lib/api-zod/src/generated/types/tag.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Universal Annotator API
 * OpenAPI spec version: 0.1.0
 */

export interface Tag {
  id: number;
  name: string;
  color: string;
  annotationCount: number;
  createdAt: Date;
}