# lib/api-zod/src/generated/types/updateAnnotationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Universal Annotator API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateAnnotationBodyType } from "./updateAnnotationBodyType";

export interface UpdateAnnotationBody {
  title?: string;
  content?: string;
  type?: UpdateAnnotationBodyType;
  color?: string;
  /** @nullable */
  sourceUrl?: string | null;
  /** @nullable */
  sourceTitle?: string | null;
  tags?: string[];
  isPinned?: boolean;
}